Add publication year and read status SVGs
I've added a new column to the table that displays
the books year of publicaiton. I'm still too add
that option to the form.

I've also changed the 'read status' column from a
yes/no to a check or cross SVG icon. I will change
the form to only give two options for this.