Validation icons and event listeners
I've changed the client side validation to display
small icons inside the inputs instead of changing
background colour.

I have also added the 'delete' button column on
the table dynamically when the function populates
I had trouble with the event listener for
dynamically created buttons but found a work
around by putting the event listener on the whole
table but testing for the target type and id.